Update to Bukkit version 1.7.2-R1
-Removed old binary files
-Updated any imported CraftBukkit 1.5.2 files to their equivalent 1.7.2
files
-Updated the health system to doubles

+++ src/net/steeleyes/catacombs/CatUtils.java
@@ -331,9 +331,9 @@ public class CatUtils {
     }
   }
   
-  public static int getThreatFixDurability(Player player, int dmg) {
+  public static double getThreatFixDurability(Player player, double dmg) {
     ItemStack stk = player.getItemInHand();
-    int threat = dmg;
+    double threat = dmg;
     if(stk==null)
       return threat;
